Extracted ActiveCards into separate file
