+++ backend/src/gamestate/mod.rs
@@ -1,3 +1,4 @@
+mod active_cards;
 pub mod game;
 pub mod player;
